// openapi-fetch-testing/src/generated/api.d.ts
export interface components {
    schemas: {
        Pet: {
            id: number;
            name: string;
            /** @enum {string} */
            species: "dog" | "cat" | "bird";
            age?: number;
        };
        CreatePetBody: {
            name: string;
            /** @enum {string} */
            species: "dog" | "cat" | "bird";
            age?: number;
        };
        UpdatePetBody: {
            name?: string;
            /** @enum {string} */
            species?: "dog" | "cat" | "bird";
            age?: number;
        };
        Error: {
            code: number;
            message: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}